Replace the existing API Documentation Doclet with a new implementation that uses the standard Doclet API available in Java 9 onward.
